requirements.txt
streamlit
pandas
pyperclip
openpyxl
python-docx